Added rate limit handling via aysnc methods
For nodes sending messages switched to use the methodsAsync call and
wait for the CompletableFuture to return before sending another message.

Not sure if this manages multiple nodes sending at the same time.

Bumped minor version number as this required no visible changes to the
user other than updating the node description XML.